feat: support twin.macro

+++ packages/toolkit/node-bundle-require/src/index.ts
@@ -1,7 +1,9 @@
 import { fs } from '@modern-js/utils';
-import { bundle, Options } from './bundle';
 
-export { bundle };
+import { bundle, Options, defaultGetOutputFile } from './bundle';
+
+export { bundle, defaultGetOutputFile };
+export type { Options };
 
 function deleteRequireCache(path: string) {
   if (require.cache[path]) {
@@ -25,7 +27,10 @@ export async function bundleRequire(filepath: string, options?: Options) {
     // The bundled file is temporary, so we should clear the require history to avoid breaking the webpack cache.
     deleteRequireCache(configFile);
   } finally {
-    fs.unlinkSync(configFile);
+    // default auto clear configFile
+    if (typeof options?.autoClear === 'undefined' || options.autoClear) {
+      fs.unlinkSync(configFile);
+    }
   }
 
   return mod;
